# backend/docs/swagger.yaml
definitions:
...
  handlers.UpdateBookingInput:
    properties:
      end_time:
        type: string
      purpose:
        type: string
      room_id:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
  models.CreateBookingInput:
    properties:
      attendees:
        type: integer
      end_time:
        type: string
      purpose:
        type: string
      room_id:
        type: string
      start_time:
        type: string
      user_email:
        type: string
      user_name:
        type: string
    required:
    - attendees
    - end_time
    - purpose
    - room_id
    - start_time
    - user_email
    - user_name
    type: object
info:
  contact: {}
paths:
...
  /api/bookings:
    get:
      consumes:
      - application/json
      description: Get list of bookings with optional pagination, room, and status
        filter
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Room ID filter
        in: query
        name: room_id
        type: string
      - description: Booking status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all bookings
      tags:
      - booking
    post:
      consumes:
      - application/json
      description: Create a new booking
      parameters:
      - description: Booking info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookingInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create booking
      tags:
      - booking
  /api/bookings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete booking
      tags:
      - booking
    get:
      consumes:
      - application/json
      description: Get detail of a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get booking detail
      tags:
      - booking
    put:
      consumes:
      - application/json
      description: Update a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Booking info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateBookingInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update booking
      tags:
      - booking
  /api/bookings/approve/{id}:
    patch:
      consumes:
      - application/json
      description: Approve a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Approve booking
      tags:
      - booking
  /api/bookings/delete/{token}:
    delete:
      consumes:
      - application/json
      description: Delete a booking using QR code token
      parameters:
      - description: QR Code Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete booking by QR token
      tags:
      - booking
  /api/bookings/reject/{id}:
    patch:
      consumes:
      - application/json
      description: Reject a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Reject booking
      tags:
      - booking
...
swagger: "2.0"